Making ReduxAssets not required for app initialization.
This is intentional change. Gaugin doesn't work with local data and gets all data from ReduxAssets provided by vangogh

To mitigate this without massive code changes in stencil we introduce DataReduxAssets a wrapper on top of vangogh provided data that implements ReduxAssets interface. This means that we'll need to move rxa to a call parameter (gaugin will have to wrap every result)

Moving rest of the icons from gaugin

Updating list styles to be closer to gaugin